requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-postgres==5.11.0
google-api-python-client
psycopg2-binary
pandas
astronomer-cosmos
dbt-postgres
apache-airflow-providers-amazon==8.17.0
emailmessage
openpyxl